feat(panel): surface work sessions as a Git page tab (route was orphaned)
