+++ panel/src/components/work-sessions/index.ts
@@ -1,2 +1,4 @@
 export { WorkSessionTable } from "./work-session-table";
 export { WorkSessionFilters } from "./work-session-filters";
+export { SessionTrendChart } from "./session-trend-chart";
+export { WorkSessionsView } from "./work-sessions-view";
